requirements.txt
anyio==3.6.2
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
fastapi==0.95.1
filelock==3.9.0
google==3.0.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
greenlet==2.0.2
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
platformdirs==2.6.2
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==1.10.7
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.15
starlette==0.26.1
tqdm==4.66.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.21.1
virtualenv==20.17.1
